% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTF_raster_plot.R
\name{ctf_heatmap}
\alias{ctf_heatmap}
\title{Create a heatplot for flow and cease-to-flow periods in a daily time-series.}
\usage{
ctf_heatmap(df, threshold = 0, units = "ML/day", format = "\%d/\%m/\%Y",
  cols = c(1, 2))
}
\arguments{
\item{df}{A dataframe with date in the first column (as character) and discharge
in the second (can be modified with cols.}

\item{threshold}{The threshold used to indicate cease to flow.}

\item{units}{The units of the time series (used for the y-axis label).}

\item{format}{The date format used (default = '\%d/\%m/\%Y').}

\item{cols}{The columns of the Date and discharge fields (default = c(1,2)).}
}
\value{
A ggplot object.
}
\description{
Create a heatplot for flow and cease-to-flow periods in a daily time-series.
}
\examples{
library(hydrostats)
data(Cooper)
ctf_heatmap(Cooper)

data(Acheron)
df <- rbind(data.frame(River='Acheron', Acheron), data.frame(River='Cooper', Cooper))
ctf_plots <- dlply(df, .(River), function(x) ctf_heatmap(x, threshold=1000, cols=c(2,3)))
ctf_plots[[1]]
ctf_plots[[2]]
}

